Add e2e tests
As part of this, it was nesscary to refactor the common test suite which was previously only used by the unit tests because Playwright's test runner uses different structual (describe/test) constructs, and requires spreading the parameters passed into the `test` method.

Thus the decision to have the common tests call a function which registers the test itself.

This also fixes a UI bug which the common tests were detecting as part of the e2e tests.

Note: Originally, the goal was to add Component Tests, but there have been technical challenges getting that to work, and given the scale of this project, the performance difference between CT and e2e tests is likely minimial.