Remove logging of "Per-pattern summaries"'s overall best/worse data Also stabilize Sword and Present display
Details:
Prior to this change, two stats were calculated (for each pattern):
1. The overall best/worst (for the pattern)
2. The per-a-metric best/worst (for the pattern)

A lot of the time, these were the same. But sometimes they diverged.
This was basically an indicator that all recommendations were not equal
(e.g. sometimes one of the recommendations would be a fox). This is not
"wrong" because we're trying to find the best recommendation for all
possible patterns, not just this pattern.

Anyway, this overall best/worst was both confusing and not even useful.
It's not useful because the overall result is pure luck; our main
concern is one of the three main target sets targeted by strats: SP SF
PF
It was also confusing because it is purely looking at the total to
uncover all titles, which results in e.g. having SP11 in the overall
best while the best for SP is actually SP[11,12] because the pattern it
picked happened to win a 50/50 pick.