1. Replaced dual-display of  Sword and Present "Per-pattern summaries" with split results 2. Rename "stepsTo" fields to be less confusing
Regarding (1), the prior display was confusing because e.g. s[1,2]+[5,6] is not saying there is s1+5 and s2+6 (it might, but it might not be) because these are semi-independent values (they are linked in that s1 and 5 would be assigned together or not at all, but you can still end up with s2+5 s1+6 which is then reported as s[1,2]+[5,6]).

By showing the found and uncovered as entirely independent values, this problem is entirely resolved.