Add new full auto-solver
This auto-solver checks all possibilities without shortcuts. The prior auto-solver (now-called short-circuit solver) only actually made moves to everything being "found" and then used math to figure out how many moves to uncover using various strategies (e.g. SwordFox, PresentFox, etc...).

This probably worked, but it was not obviously correct. This new approach is much simpler, and thus it is much easier to conclude it is also correct. The main downside is it is much slower to run because not only do we need to go through more possibilities (like every possible permutation of foxes, which is often 4!) for each pattern, we have to evaluate each pattern multiple times for different possible strategies.